Re-organise and extend path loading logic to make it easier to run integration tests using Delve/GoLand
