+++ internal/test/paths/BUILD.bazel
@@ -2,12 +2,15 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["bazel.go"],
+    srcs = [
+        "bazel.go",
+        "paths.go",
+    ],
     data = [
         "//deploy/crds:crds.yaml",
     ],
-    importpath = "github.com/jetstack/cert-manager/pkg/api/testing",
-    visibility = ["//visibility:public"],
+    importpath = "github.com/jetstack/cert-manager/internal/test/paths",
+    visibility = ["//:__subpackages__"],
 )
 
 filegroup(
