+++ test/integration/framework/BUILD.bazel
@@ -11,8 +11,8 @@ go_library(
     deps = [
         "//cmd/webhook/app:go_default_library",
         "//cmd/webhook/app/testing:go_default_library",
+        "//internal/test/paths:go_default_library",
         "//pkg/api:go_default_library",
-        "//pkg/api/testing:go_default_library",
         "//pkg/client/clientset/versioned:go_default_library",
         "//pkg/client/informers/externalversions:go_default_library",
         "//pkg/controller:go_default_library",
